Update Dockerfile and exec_all_pipelines.sh script
Updated `Dockerfile` to include the installations of necessary packages, the `owltools` download and setting its permissions. Also, `exec_all_pipelines.sh` script has been modified to streamline the folder creation process and to correct the paths to the project files. The changes implemented aim to ensure smooth deployment and operation of the Wormbase project container.